Legacy LORA Radio Node V1 repo/room-sensors-master/room-sensors-master/webapplication/requirements.txt
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
paho-mqtt==1.3.1
pytz==2018.5
gunicorn
